# kabtak/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "CheckAccepted": {
        "properties": {
          "check_id": {
            "title": "Check Id",
            "type": "string"
          },
          "poll_after_ms": {
            "default": 2000,
            "title": "Poll After Ms",
            "type": "integer"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "status": {
            "default": "queued",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "check_id",
          "run_id"
        ],
        "title": "CheckAccepted",
        "type": "object"
      },
...
      "ExampleRead": {
        "properties": {
          "academic_year": {
            "title": "Academic Year",
            "type": "string"
          },
          "application_type": {
            "title": "Application Type",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "expected_student_deadline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Student Deadline"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "programme_name": {
            "title": "Programme Name",
            "type": "string"
          },
          "reference_time": {
            "format": "date-time",
            "title": "Reference Time",
            "type": "string"
          },
          "source_url": {
            "title": "Source Url",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "description",
          "programme_name",
          "academic_year",
          "application_type",
          "reference_time",
          "expected_student_deadline",
          "source_url"
        ],
        "title": "ExampleRead",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Deadline and evidence checking API.",
    "title": "Kabtak API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/examples": {
      "get": {
        "operationId": "list_examples_v1_examples_get",
        "parameters": [
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExampleRead"
                  },
                  "title": "Response List Examples V1 Examples Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Examples",
        "tags": [
          "examples"
        ]
      }
    },
    "/v1/examples/{example_id}/replay": {
      "post": {
        "operationId": "replay_example_v1_examples__example_id__replay_post",
        "parameters": [
          {
            "in": "path",
            "name": "example_id",
            "required": true,
            "schema": {
              "title": "Example Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          },
          {
            "in": "header",
            "name": "x-internal-token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Internal-Token"
            }
          }
        ],
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CheckAccepted"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Replay Example",
        "tags": [
          "examples"
        ]
      }
    },
...
  }
}